refactor: extract shell helpers to shell-utils.js and add changeset (issue #91)
Extracted SHELL_NAMES, isInteractiveShellCommand, isShellInvocationWithArgs,
buildShellWithArgsCmdArgs, and buildDisplayCommand from isolation.js into a
new shell-utils.js module to keep isolation.js under the 1000-line CI limit
(984 lines after extraction, down from 1009).

Added changeset file for the patch release.
